backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
install scipyscikit-learn==1.3.0
joblib==1.3.2
mysql-connector-python==8.0.33
PyJWT==2.8.0
Werkzeug==2.3.7
numpy==1.26.4
pip ==1.11.4
scikit-learn==1.3.2
pandas==2.0.3